+++ src/tests/__fixtures__/ComponentClass.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { createKeydownFromClick } from '../../api/createKeydownFromClick'
 import { Options } from '../../typings'
-import { DATA_TESTID } from '../utils'
+import { DATA_TEST_ID } from '../utils'
 
 export type ComponentClassProps = {
   onClick: React.MouseEventHandler
@@ -10,7 +10,6 @@ export type ComponentClassProps = {
 }
 
 export class ComponentClass extends React.Component<ComponentClassProps> {
-  private handleClick: React.MouseEventHandler
   private handleKeyDown: React.KeyboardEventHandler
 
   public constructor(props: ComponentClassProps) {
@@ -18,22 +17,16 @@ export class ComponentClass extends React.Component<ComponentClassProps> {
 
     const { onClick, options } = props
 
-    this.handleClick = onClick
-
-    this.handleKeyDown = createKeydownFromClick(this.handleClick, options)
+    this.handleKeyDown = createKeydownFromClick(onClick, options)
   }
 
   public render(): React.ReactNode {
     return (
-      <div
-        data-testid={DATA_TESTID}
-        onClick={this.handleClick}
+      <button
+        data-testid={DATA_TEST_ID}
+        onClick={this.props.onClick}
         onKeyDown={this.handleKeyDown}
-        role="button"
-        tabIndex={0}
-      >
-        lala
-      </div>
+      />
     )
   }
 }