+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `shouldPropagate` option, which controls whether the event should propagate; it defaults to `true`.
+
 ## [2.0.0] - 2020-03-06
 
 ### Changed